requirements.txt
torch>=2.2
torchvision>=0.17
albumentations>=1.4.0
opencv-python>=4.8
numpy
tqdm
pyyaml
scikit-image
tensorboard
